+++ vortex-file/src/v2/writer.rs
@@ -9,7 +9,7 @@ use vortex_io::VortexWrite;
 use vortex_layout::strategies::LayoutStrategy;
 
 use crate::v2::footer::{FileLayout, Postscript, Segment};
-use crate::v2::segments::BufferedSegmentWriter;
+use crate::v2::segments::writer::BufferedSegmentWriter;
 use crate::v2::strategy::VortexLayoutStrategy;
 use crate::{EOF_SIZE, MAGIC_BYTES, MAX_FOOTER_SIZE, VERSION};
